# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TICKET API",
        "description": "Documentation de mon API Laravel",
        "contact": {
            "name": "Support API",
            "email": "shadracmeshacdimon@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/ticket-types": {
            "get": {
                "tags": [
                    "TicketTypes"
                ],
                "summary": "Obtenir tous les types de tickets",
                "operationId": "64e4b205a82da5044837539a6664d170",
                "responses": {
                    "200": {
                        "description": "Liste des types de tickets",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TicketType"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "TicketTypes"
                ],
                "summary": "Créer un nouveau type de ticket",
                "operationId": "a0422b6de44e07976c5380641c187507",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "ticket_type_event_id",
                                    "ticket_type_name",
                                    "ticket_type_price",
                                    "ticket_type_quantity",
                                    "ticket_type_real_quantity",
                                    "ticket_type_total_quantity",
                                    "ticket_type_description"
                                ],
                                "properties": {
                                    "ticket_type_event_id": {
                                        "type": "integer"
                                    },
                                    "ticket_type_name": {
                                        "type": "string",
                                        "maxLength": 50
                                    },
                                    "ticket_type_price": {
                                        "type": "integer"
                                    },
                                    "ticket_type_quantity": {
                                        "type": "integer"
                                    },
                                    "ticket_type_real_quantity": {
                                        "type": "integer"
                                    },
                                    "ticket_type_total_quantity": {
                                        "type": "integer"
                                    },
                                    "ticket_type_description": {
                                        "type": "string",
                                        "maxLength": 16777215
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Type de ticket créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TicketType"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/events/{eventId}/ticket-types": {
            "get": {
                "tags": [
                    "TicketTypes"
                ],
                "summary": "Obtenir tous les types de tickets pour un événement spécifique",
                "operationId": "d913cf056cf85e181d64f563e15314d3",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des types de tickets pour l'événement",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TicketType"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun type de ticket trouvé pour cet événement",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "event_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Event": {
                "required": [
                    "event_id",
                    "event_title",
                    "event_description",
                    "event_date",
                    "event_image",
                    "event_category",
                    "event_city",
                    "event_address",
                    "event_status"
                ],
                "properties": {
                    "event_id": {
                        "type": "integer"
                    },
                    "event_title": {
                        "type": "string",
                        "maxLength": 30
                    },
                    "event_description": {
                        "type": "string",
                        "maxLength": 16777215
                    },
                    "event_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "event_image": {
                        "type": "string",
                        "maxLength": 200
                    },
                    "event_category": {
                        "type": "string",
                        "enum": [
                            "Autre",
                            "Concert-Spectacle",
                            "Diner Gala",
                            "Festival",
                            "Formation"
                        ]
                    },
                    "event_city": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "event_address": {
                        "type": "string",
                        "maxLength": 200
                    },
                    "event_status": {
                        "type": "string",
                        "enum": [
                            "upcoming",
                            "completed",
                            "cancelled"
                        ]
                    }
                },
                "type": "object"
            },
...
            "TicketType": {
                "required": [
                    "ticket_type_event_id",
                    "ticket_type_name",
                    "ticket_type_price",
                    "ticket_type_quantity",
                    "ticket_type_real_quantity",
                    "ticket_type_total_quantity",
                    "ticket_type_description"
                ],
                "properties": {
                    "ticket_type_id": {
                        "description": "ID du type de ticket",
                        "type": "integer"
                    },
                    "ticket_type_event_id": {
                        "description": "ID de l'événement associé",
                        "type": "integer"
                    },
                    "ticket_type_name": {
                        "description": "Nom du type de ticket",
                        "type": "string",
                        "maxLength": 50
                    },
                    "ticket_type_price": {
                        "description": "Prix du type de ticket",
                        "type": "integer"
                    },
                    "ticket_type_quantity": {
                        "description": "Quantité disponible du type de ticket",
                        "type": "integer"
                    },
                    "ticket_type_real_quantity": {
                        "description": "Quantité réelle du type de ticket",
                        "type": "integer"
                    },
                    "ticket_type_total_quantity": {
                        "description": "Quantité totale du type de ticket",
                        "type": "integer"
                    },
                    "ticket_type_description": {
                        "description": "Description du type de ticket",
                        "type": "string",
                        "maxLength": 16777215
                    },
                    "event": {
                        "$ref": "#/components/schemas/Event"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Use the token provided after logging in.",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}